backend/requirements.txt
fastapi
uvicorn
openai
langchain
chromadb
unstructured==0.7.12
tiktoken
langchain-community
fastapi-cors
fastapi-socketio
langchain-openai
python-dotenv
langchain_huggingface
accelerate
transformers
gradio 
xformer
bitsandbytes 
unstructured 
sentence-transformers
aiofiles